I forgot that when making a mutation there needsto be knowledge about all the neurons in the network, not just the connections. Since the genome only contains info about the connections, I had to add a few specific genes with a innovation of -1 and type NEURON_TYPE_INPUT/OUTPUT to specify that a neuron exists without a connection

+++ src/network.cpp
@@ -12,7 +12,7 @@ void AxonNeuronPair::setNeuron(Neuron *ner){
 }
 
 float AxonNeuronPair::getValue(){
-    return (neuron->getValue() * weight * enabled);
+    return (float)(neuron->getValue() * weight * enabled);
 }
 
 Neuron::Neuron(int i){
@@ -29,7 +29,7 @@ float Neuron::getValue(){
     if(type == 0){
         return result;
     }else if(type == 1){
-        for(int i = 0; i < inputs.size(); i++){
+        for(int i = 0; i < (int) inputs.size(); i++){
             input += inputs.at(i)->getValue();
         }
         return input;
@@ -50,7 +50,7 @@ void Network::update(){
         return;
     }
     outputs.clear();
-    for(int i = 0; i < outputNeurons.size(); i++){
+    for(int i = 0; i < (int) outputNeurons.size(); i++){
        outputs.push_back(outputNeurons.at(i)->getValue()); 
     }
 }
